TODO: work out why there is a discrepancy in the VecFullTest output with the previous commit.
This commit papers over the problem (by updating the expected results) rather
than fixing it because I don't have time to do so now. The StreamValidator
points out a difference in random number generation with the original
change-set, but when extra validation is added to the random number code before
and after the change-set the location of the first difference moves. See
comments in Vaccine.cpp.

The best solution long term is probably to switch to a memory-safe programming
language...