Add XML and hooks for configurable treatments; incomplete implementation.
This updates the XML as needed for configurable treatments for MDA (except that
it may be better eventually only to reference the treatment type from the MDA
description and configure it externally).
It also adds hooks to configure and use configured treatments, but these don't
work yet.
It breaks any tests using the old MDA description.