Partial compatibility with C++11 (hacky, no 1-day TS CM)
This makes OpenMalaria build with C++11.

Caveat: 1-day case management and thus all 1-day timestep models are disabled
since the configuration parser needs tweaking (in ways I have yet to discover).

As such, the following tests are expected to fail:
    unittest
    1DayIntervs.xml
    Cohort.xml
    EffectiveDrug.xml
    Empirical.xml
    ESCMTest.xml
    ESTS.xml
    Molineaux.xml
    MSAT.xml
    Penny.xml

Caveat: there are a lot of warnings about the deprecated auto_ptr<T>. This
could probably largely be replaced with unique_ptr<> but should be reviewed.