Update tests to simplify usage of 1-day CM decision trees
Remove usage of the "select range" schedule modifier so that this can be
removed from code in the interests of simplicity.
Simplify some decision trees by removing unnecessary logic.
These changes are not completely verifiable as having the same effect, but
differences in output shrink at a population size of 100'000.