Larval competition: implement "option 2" of proposal, detailed below.
Instead of using a single "resources" number for each species, include weighting factors. So instead of

    gamma bar * (sum_j y_j)

we use

    sum_j gamma_j * y_j

in the emergence equation (where gamma is the larval resources and y is the number of juvenile stages, j is the vector species).

I'm not sure this will work, but it at least avoids trying to fit multiple output numbers from a single input.